fix delete route

+++ src/routes/Mangas/MangaController.ts
@@ -72,6 +72,14 @@ export const deleteManga = async (req: Request, res: Response) => {
   if (!MangaID) return res.status(400).send({ error: 'Missing id parameter' })
 
   try {
+    const manga = await prisma.manga.findUnique({
+      where: {
+        MangaID,
+      },
+    })
+
+    if (!manga) return res.status(404).send({ error: 'Manga doesnt exist' })
+
     await prisma.manga.delete({
       where: {
         MangaID,
